requirements.txt
eth-brownie
scipy
matplotlib